feat: complete handshake between slave and master
